Validate HTTP response codes and several new endpoints

+++ src/comment.hpp
@@ -6,9 +6,9 @@ class RuqqusComment {
         std::string author;
         std::string body;
         std::string body_html;
-        uintmax_t created_utc;
-        uintmax_t downvotes;
-        uintmax_t edited_utc;
+        int created_utc;
+        int downvotes;
+        int edited_utc;
         std::string fullname;
         std::string guild_name;
         std::string id;
@@ -18,13 +18,13 @@ class RuqqusComment {
         bool is_nsfl;
         bool is_nsfw;
         bool is_offensive;
-        uintmax_t level;
+        int level;
         std::string parent;
         std::string permalink;
         std::string post;
-        uintmax_t score;
+        int score;
         std::string title;
-        uintmax_t upvotes;
+        int upvotes;
 };
 
 #endif // COMMENT_HPP_INCLUDED
